Dockerise application and migrate to MySQL

+++ db/migrate/20190112091319_create_beatmaps.rb
@@ -2,10 +2,10 @@ class CreateBeatmaps < ActiveRecord::Migration[5.2]
   def change
     create_table :beatmaps do |t|
       t.string :name
-      t.integer :online_id
+      t.bigint :online_id
       t.float :star_difficulty
       t.string :difficulty_name
-      t.integer :max_combo
+      t.bigint :max_combo
 
       t.timestamps
     end
